feat: create points of route movie create, findById, findAll
