+++ src/main/java/com/flav/trailers/commons/ApplicationConfig.java
@@ -1,0 +1,15 @@
+package com.flav.trailers.commons;
+
+import org.modelmapper.ModelMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class ApplicationConfig {
+
+    @Bean
+    public ModelMapper modelMapper() {
+        return new ModelMapper();
+    }
+
+}
