+++ src/main/java/com/flav/trailers/context/trailers/genders/application/FindAllGenderUseCase.java
@@ -1,0 +1,20 @@
+package com.flav.trailers.context.trailers.genders.application;
+
+import com.flav.trailers.context.trailers.genders.domain.models.Gender;
+import com.flav.trailers.context.trailers.genders.domain.repositories.IGenderCRUDRepository;
+
+import java.util.List;
+
+public class FindAllGenderUseCase {
+
+    private final IGenderCRUDRepository repo;
+
+    public FindAllGenderUseCase(IGenderCRUDRepository repo) {
+        this.repo = repo;
+    }
+
+    public List<Gender> run() {
+        return  repo.findAll();
+    }
+
+}
