+++ src/main/java/com/flav/trailers/context/trailers/genders/domain/constants/GenderConstants.java
@@ -1,0 +1,8 @@
+package com.flav.trailers.context.trailers.genders.domain.constants;
+
+public class GenderConstants {
+
+    public static final String GENDER_NOT_FOUND = "Gender not found by id: %s";
+    public static final String GENDER_EXITS = "Gender name (%s) already exist";
+
+}
