+++ src/main/java/com/flav/trailers/context/trailers/genders/domain/exeptions/GenderResourceExist.java
@@ -1,0 +1,20 @@
+package com.flav.trailers.context.trailers.genders.domain.exeptions;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class GenderResourceExist extends RuntimeException {
+
+    private String message;
+    private HttpStatus status;
+
+    public GenderResourceExist(String message, HttpStatus status) {
+        this.message = message;
+        this.status = status;
+    }
+}
