+++ src/main/java/com/flav/trailers/context/trailers/genders/domain/models/Gender.java
@@ -1,0 +1,15 @@
+package com.flav.trailers.context.trailers.genders.domain.models;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Gender {
+
+    private Long id;
+    private String name;
+
+}
