+++ src/main/java/com/flav/trailers/context/trailers/movies/domain/constants/MovieConstants.java
@@ -1,0 +1,8 @@
+package com.flav.trailers.context.trailers.movies.domain.constants;
+
+public class MovieConstants {
+
+    public static final String MOVIE_NOT_FOUND = "Movie not found by id: %s";
+    public static final String MOVIE_EXITS = "Movie name (%s) already exist";
+
+}
