+++ src/main/java/com/flav/trailers/context/trailers/movies/infraestructure/persistence/entities/MovieEntity.java
@@ -1,0 +1,27 @@
+package com.flav.trailers.context.trailers.movies.infraestructure.persistence.entities;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.Date;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+@Entity
+@Table(name = "movies")
+public class MovieEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+    private String description;
+    private String image;
+    private String video;
+    @Column(name = "release_date")
+    private Date releaseDate;
+
+}
