--- src/main/java/com/flav/trailers/genders/infraestructure/persistence/entities/GenderEntity.java
@@ -1,20 +1,0 @@
-package com.flav.trailers.genders.infraestructure.persistence.entities;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Entity
-@Table(name = "gender")
-public class GenderEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-
-}
